pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "twin_crew"
version = "0.1.0"
description = "Enrique Crew AI System - Multi-agent scheduling assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai==0.100.0",
    "langchain-openai>=0.3.33",
    "pydantic-settings>=2.10.1",
    "pytz>=2025.2",
]

[project.scripts]
twin_crew = "twin_crew.main:run"
run_crew = "twin_crew.main:run"
train = "twin_crew.main:train"
replay = "twin_crew.main:replay"
test = "twin_crew.main:test"
chat = "twin_crew.main:chat"

[tool.hatch.build.targets.wheel]
packages = ["src/twin_crew"]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.13.0",
    "mypy>=1.18.1",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["crew"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "crewai.*",
    "langchain.*",
    "pytz.*",
]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
